fix build script
