+++ save-preprocessor/build.gradle.kts
@@ -4,7 +4,7 @@ import com.saveourtool.save.buildutils.configureSpotless
 plugins {
     id("com.saveourtool.save.buildutils.kotlin-jvm-configuration")
     alias(libs.plugins.kotlin.plugin.serialization)
-    id("com.saveourtool.save.buildutils.spring-boot-configuration")
+    id("com.saveourtool.save.buildutils.spring-boot-app-configuration")
 }
 
 dependencies {
